HHH-10398 Allow MOD column naming to be driven by a strategy
In the past the MOD columns were constructed based on the property name,
therefore if users specified a @Column/@JoinColumn like annotation and
changed the underlying schema column, the MOD column would continue to
be derived based on the property name.

This enhancement introduces a new ModifiedColumnNamingStrategy SPI that
comes with two implementations, a default/legacy mode that maintains
the prior naming model and an improved mode that will derive the MOD
name based on the naming strategy ORM used to derive the column name.